Fornt-End/font-end-app/src/app/shared/input-markdown/input-markdown.component.ts
import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';

@Component({
  selector: 'app-input-markdown',
  templateUrl: './input-markdown.component.html',
  styleUrls: ['./input-markdown.component.css']
})
export class InputMarkdownComponent implements OnInit {

  @Input()
  contentMarkdown = '';

  @Input()
  placeHolderTextArea: string = 'Texto';

  @Output()
  changeMarkdown: EventEmitter<string> = new EventEmitter<string>();

  constructor() { }

  ngOnInit(): void {
  }

  onInputChange(event: any) {
    const value = (event.target as HTMLTextAreaElement)?.value;
    this.changeMarkdown.emit(value);
  }
}
